+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2025-05-12
+
+### Added
+
+- Added program rules schemas and program template JSON blobs
+
+### Changed
+
+- BREAKING: updated `json-schema-to-typescript@10.1.5` to
+  `json-schema-to-typescript-with-deps@0.0.1` (based on `json-schema-to-typescript@15`)
+- BREAKING: dropped support for Node versions older than 18 during build process (does
+  not affect generated & published schemas)
+
 ## [1.28.9] - 2025-02-04
 
 ### Changed
@@ -187,6 +200,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Removed "pinterestShareBody" as a required field for the "PINTEREST" share medium
   - Removed "pinterestImageURL" as a required field for the "PINTEREST" share medium
 
+[2.0.0]: https://github.com/saasquatch/schema/releases/tag/v2.0.0
 [1.28.9]: https://github.com/saasquatch/schema/releases/tag/v1.28.9
 [1.28.8]: https://github.com/saasquatch/schema/releases/tag/v1.28.8
 [1.28.7]: https://github.com/saasquatch/schema/releases/tag/v1.28.7